Include the error in the Objective-C selector for throwing methods.
When computing an Objective-C selector for a throwing method,
introduce an "error:" selector piece at the location specified by the
foreign error convention. If it's part of the first selector piece,
append "WithError" to the method base name.

Also require the manually-specified @objc(selector) of a throwing
method to include a selector piece for the error parameter.

Swift SVN r27276