Allow an explicit @objc(selector) to indicate where the error parameter goes.
For a throwing @objc method, we place the error parameter just before
the last parameter of (possibly-optional) function type. Allow this
default to be overridden with an explicit @objc(selector) attribute,
which can contain an "error" selector piece to indicate where the
error parameter goes (or an "Error" word suffix on the first selector
piece to put the error argument at the beginning).

Swift SVN r27278