fix <rdar://problem/20608881> DI miscompiles this testcase into a memory leak
When emiting logic for conditional destruction of a variable, we weren't considering
a delegated init as a store to the variable that needed to be tracked.  This caused us
to miscompile situations like this testcase where the self.init was conditional:

struct X {
  var c : C
  init() { c = C()}
  init?(a : Bool) {
    if a { self.init() }
    return nil
  }
}

This was exposed by other work, but needs to be fixed in any case.



Swift SVN r27471