The retain and release counts of a partial_apply may be unbalanced.
In hasOnlyRetainReleaseUsers() we both assert that the retain/release
are balanced, as well as do a normal check and bail.

The assert is overzealous since we can legitimately have cases where the
releases are removed due to being unreachable. In this case we should
just do the normal check and bail.

Fixes rdar://problem/20565974.

Swift SVN r27476