Remove some (ultimately) no-op code from SILGen of RebindSelfInConstructorExpr: it would drop in a dealloc_box or strong_release after rebinding, which is incorrect (the box is used for any loads of self after the rebind!) but also pointless: the AllocBoxToStack pass saved the day by promoting the box to a stack address and zapping these instructions.
Interestingly, AllocBoxToStack *always* succeeds in class initializers, because the
self value is treated as a let value (even though in a box) and captured by value,
therefore there is no way to block the stack promotion that I can find.

In any case, this code is in my way and not doing anything, buhbye.




Swift SVN r27517