Reenable some devirtualization cases disabled in r26152.
In r26152 a crash was fixed that had been introduced by previous
improvements, but as a result of fixing the crash some cases where we
previously devirtualized were disabled with the intent that they would
be reenabled after further refactoring work was completed.

This commit restores that functionality and reenables the tests.

Swift SVN r27575