Reapply "SimplifyCFG: Fix a bug in the jump threading code"
This reverts commit r27739 reapplying r27722.

The test (validation/stdlib/Hashing.swift) that failed is expected to sometimes
fails.

Original message

"SimplifyCFG: Fix a bug in the jump threading code

When jump threading a block we used to propagate phi values directly into the
threaded block - instead of leaving in a copy. Because of this the SSA updater
would propagate the value feeding the copy from the next iteration.
Now when jump threading the destination block into the edge we leave in the phi
(copy) such that the SSA updater picks up value of the right iteration.

rdar://20617338"

Swift SVN r27741