Teach tryDeleteDeadClosure how to properly delete all types of captured arguments.
Previously, we relied on certain types not being passed in by the users of the
given function. Since I had a lot on my plate on the time and I knew it was safe
I left it in. Now I am going to be using this function in more places so it
needs to handle everything correctly.

rdar://19552593

Swift SVN r27748